requirements.txt
numpy
Pillow
numpy-stl
scipy
flask
